refactor(blorktools): animation directory
