+++ packages/blorktools/src/asset_debugger/core/models.js
@@ -630,7 +630,7 @@ function checkAndApplyCustomDisplaySettings() {
             import('../core/glb-utils.js'),
             import('../core/string-serder.js'),
             import('../core/texture-util.js'),
-            import('../core/css3d-util.js'),
+            import('./animation/css3d-util.js'),
             import('../ui/scripts/model-integration.js') // Add model integration import
         ]).then(([glbUtils, stringSerder, textureUtil, css3dUtil, modelIntegration]) => {
             // If no meshes, resolve immediately
